requirements.txt
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
dotenv==0.9.9
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
idna==3.11
numpy==2.3.4
pandas==2.3.3
proto-plus==1.26.1
protobuf==6.33.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
